+++ app/_interfaces/FormInputData.ts
@@ -1,8 +1,6 @@
-import { FieldError, FieldValues, Path, RegisterOptions } from "react-hook-form";
+import { FieldValues, Path, RegisterOptions } from "react-hook-form";
 
 export interface FormInputData<TFormData extends FieldValues>{
-  // エラーメッセージ
-  error?: FieldError;
   // ラベル表示文字
   labelText: string;
   // inputType
